requirements.txt
pyjson5
pillow
numpy
matplotlib
tensorflow
regex
pandas
keras
opencv-python
